+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* Fix type exception in addDefFactory2.
+* LifecycleOwner adds extension getViewModelProviderByApp getViewModelProviderByRoute.
+
 ## 1.0.0
 
 * Add a new factory based on lifecycle